+++ Mods ChoGGi/Map Images Pack/metadata.lua
@@ -1,8 +1,9 @@
 return PlaceObj("ModDef", {
+--~ 	"title", "Map Images Pack v0.1",
 	"title", "Map Images Pack",
 	"version", 20,
 	"version_major", 0,
-	"version_minor", 1,
+	"version_minor", 2,
 	"saved", 1542888000,
 	"image", "Preview.png",
 	"id", "ChoGGi_MapImagesPack",