+++ Mods ChoGGi/RC Mechanic/Code/Script.lua
@@ -5,8 +5,10 @@ local description = [[Give me your tired, your poor,
 Your huddled masses yearning to breathe free,
 The wretched refuse of your teeming shore.]]
 local display_icon = CurrentModPath .. "UI/rover_combat.png"
-local idle_text = _InternalTranslate(T(6722--[[Idle--]]))
-local travel_text = _InternalTranslate(T(63--[[Travelling--]]))
+local idle_text = ChoGGi.ComFuncs.Translate(6722--[[Idle--]])
+local travel_text = ChoGGi.ComFuncs.Translate(63--[[Travelling--]])
+
+GlobalVar("g_RCMechanicRepairing",{})
 
 DefineClass.RCMechanic = {
 	__parents = {
@@ -17,6 +19,10 @@ DefineClass.RCMechanic = {
 	description = description,
 	display_icon = display_icon,
 	display_name = name,
+
+	going_to_repair = false,
+	blinky = false,
+	repairing_list = false,
 
 	entity = "CombatRover",
 	accumulate_dust = false,
@@ -34,7 +40,21 @@ DefineClass.RCMechanicBuilding = {
 }
 
 function RCMechanic:GameInit()
---~ 	BaseRover.GameInit(self)
+	-- add a blinky
+	self.blinky = RotatyThing:new()
+	self.blinky:SetVisible()
+
+	self:Attach(self.blinky)
+
+	-- move blinky above bar thingy
+	local offset = self:GetVisualPos() - MovePointAway(
+		self:GetSpotLoc(self:GetSpotBeginIndex("Origin")),
+		self:GetSpotLoc(self:GetSpotBeginIndex("Particle1")),
+		200
+	)
+	self.blinky:SetAttachOffset(
+		point(offset:x(),offset:y(),self:GetObjectBBox():sizez())
+	)
 
 	-- select sounds
 	self.fx_actor_class = "AttackRover"
@@ -50,9 +70,7 @@ end
 
 --~ function RCMechanic:GetStatusUpdate()
 function RCMechanic:Getui_command()
---~ 	local info =
---~ 	info[1] = "show what it's doing..."
-	return table.concat({self.status_text}, "<newline><left>")
+	return ChoGGi.ComFuncs.TableConcat({self.status_text}, "<newline><left>")
 end
 
 function RCMechanic:GotoFromUser(...)
@@ -62,7 +80,10 @@ end
 
 -- for auto mode
 function RCMechanic:ProcAutomation()
+	self.repairing_list = self.repairing_list or g_RCMechanicRepairing or {}
+
 	local unreachable_objects = self:GetUnreachableObjectsTable()
+	unreachable_objects = unreachable_objects or {}
 
 	local rover = MapFindNearest(self, "map", "BaseRover", "Drone" ,function(o)
 		local go_fix_it
@@ -103,13 +124,18 @@ function RCMechanic:ProcAutomation()
 		if go_fix_it then
 			return not unreachable_objects[o]
 		end
---~ 		if o.command_centers and #o.command_centers == 0 and o:IsMalfunctioned() and o.repair_work_request:CanAssignUnit() then
---~ 			return not unreachable_objects[o]
---~ 		end
 	end)
 
 	if rover then
-		local pos = GetPassablePointNearby(rover:GetPos())
+		local visual_pos = rover:GetVisualPos()
+		-- don't go if someone else is on the job
+		local string_pos = tostring(visual_pos)
+		if self.repairing_list[string_pos] then
+			return 10000
+		end
+		self.repairing_list[string_pos] = true
+
+		local pos = GetPassablePointNearby(visual_pos)
 		if self:HasPath(pos, "Workrover") then
 			self.status_text = [[Was the dark of the moon on the sixth of June
 In a <color 199 124 45>]] .. (rover.name ~= "" and rover.name or rover.class) .. [[</color> pullin' logs
@@ -119,9 +145,14 @@ We is headin' for bear on I-one-oh
 'Bout a mile outta Shaky Town
 I says, "Pig Pen, this here's the Rubber Duck.
 "And I'm about to put the hammer down."]]
+
+			-- add a pretty light
+			self.blinky:SetVisible(true)
+			self.going_to_repair = string_pos
+			self.move_speed = 2 * self.base_move_speed
+			-- and we're off
 			self:Goto(pos)
-			-- find a way to slow this down?
-			rover:CheatCleanAndFix()
+			rover:RCMech_CleanAndFix()
 
 			self.status_text = idle_text
 			return 5000
@@ -135,6 +166,13 @@ I says, "Pig Pen, this here's the Rubber Duck.
 end
 
 function RCMechanic:Idle()
+	if self.going_to_repair then
+		self.blinky:SetVisible(false)
+		self.move_speed = self.base_move_speed
+		self.repairing_list[self.going_to_repair] = nil
+		self.going_to_repair = false
+	end
+
 	self.status_text = idle_text
 	local sleep = self:ProcAutomation()
 	self:SetState("idle")
@@ -170,72 +208,77 @@ function OnMsg.ClassesPostprocess()
 	end
 end
 
---~ -- add some prod info to selection panel
---~ function OnMsg.ClassesBuilt()
---~ 	local rover = XTemplates.ipRover[1]
-
---~ 	-- check for and remove existing template
---~ 	ChoGGi.ComFuncs.RemoveXTemplateSections(rover,"ChoGGi_Template_RCMechanic_Prod",true)
-
---~ 	-- we replace status
---~ 	local status = table.find(rover, "Icon", "UI/Icons/Sections/sensor.tga")
---~ 	if status then
---~ 		rover[status]:delete()
---~ 		table.remove(rover,status)
---~ 	else
---~ 		-- screw it stick it at the end
---~ 		status = #rover
---~ 	end
-
---~ 	table.insert(
---~ 		rover,
---~ 		#rover,
---~ 		PlaceObj('XTemplateTemplate', {
---~ 			"ChoGGi_Template_RCMechanic_Prod", true,
---~ 			"__context_of_kind", "RCMechanic",
---~ 			"__template", "InfopanelActiveSection",
---~ 			"Title", T(6924, "Repair"),
---~ 			"Icon", "UI/Icons/Sections/construction.tga",
---~ 		}, {
---~ 			PlaceObj("XTemplateTemplate", {
---~ 				"__template", "InfopanelText",
---~ 				"Text",  T(0,"<ui_command>"),
---~ 			}),
---~ 		})
---~ 	)
---~ end
-
-
--- CheatCleanAndFix
 local function CheatAddDust(self)
+	-- get dust amount, and convert to percentage
+	local dust_amt = (obj:GetDust() + 0.0) / 100
+	if dust_amt ~= 0.0 then
+		local value = 100
+		sprinkler:ForEachAttach(SprinklerColour,-8249088)
+		while true do
+			if value == 0 then
+				break
+			end
+			value = value - 1
+			obj:SetDust(dust_amt * value, DustMaterialExterior)
+			Sleep(100)
+		end
+		sprinkler:ForEachAttach(SprinklerColour,-10197916)
+	end
+end
+
+-- RCMech_CleanAndFix
+local DustMaterialExterior = const.DustMaterialExterior
+local function RCMech_ResetDust(self)
 	self.dust = self:GetDustMax()-1
 	self:SetDustVisuals()
-end
-Drone.CheatAddDust = CheatAddDust
-BaseRover.CheatAddDust = CheatAddDust
+	Sleep(10)
 
-Drone.CheatCleanAndFix = function(self)
-	CreateRealTimeThread(function()
+	-- get dust amount, and convert to percentage
+	local dust_amt = (self:GetDust() + 0.0) / 100
+	if dust_amt ~= 0.0 then
+		local value = 100
+		while true do
+			if value == 0 or not IsValid(self) then
+				break
+			end
+			value = value - 1
+			self:SetDust(dust_amt * value, DustMaterialExterior)
+			Sleep(50)
+		end
+	end
+end
+Drone.RCMech_ResetDust = RCMech_ResetDust
+BaseRover.RCMech_ResetDust = RCMech_ResetDust
+
+Drone.RCMech_CleanAndFix = function(self)
+	CreateGameTimeThread(function()
+		if not IsValid(self) then
+			return
+		end
 		self.auto_connect = false
 		if self.malfunction_end_state then
 			self:PlayState(self.malfunction_end_state, 1)
-			if not IsValid(self) then
-				return
-			end
 		end
-		self:CheatAddDust()
-		Sleep(10)
-		self.dust = 0
-		self:SetDustVisuals()
+
+		self:RCMech_ResetDust()
+		if self.command == "NoBattery" then
+			self.battery = self.battery_max
+			self:SetCommand("Fixed", "noBatteryFixed")
+    elseif self.command == "Malfunction" or self.command == "Freeze" and self:CanBeThawed() then
+			self:SetCommand("Fixed", "breakDownFixed")
+		else
+			self:SetCommand("Fixed", "Something")
+		end
+
 		RebuildInfopanel(self)
  end)
 end
-BaseRover.CheatCleanAndFix = function(self)
-	CreateRealTimeThread(function()
-		self:CheatAddDust()
-		Sleep(10)
-		self.dust = 0
-		self:SetDustVisuals()
+BaseRover.RCMech_CleanAndFix = function(self)
+	CreateGameTimeThread(function()
+		if not IsValid(self) then
+			return
+		end
+		self:RCMech_ResetDust()
 		self:Repair()
  end)
 end