+++ Mods ChoGGi/View Colony Map/metadata.lua
@@ -13,23 +13,26 @@ return PlaceObj("ModDef", {
 			"version_minor", 1,
 		}),
 	},
+--~ 	"title", "View Colony Map v0.9",
 	"title", "View Colony Map",
 	"version", 20,
-	"version_major", 0,
-	"version_minor", 9,
-	"saved", 1542888000,
+	"version_major", 1,
+	"version_minor", 0,
+	"saved", 1555416000,
 	"image", "Preview.png",
 	"id", "ChoGGi_ViewColonyMap",
 	"steam_id", "1491973763",
 	"pops_any_uuid", "28b23a4f-7e8f-49b0-965a-3c14a8e4b919",
 	"author", "ChoGGi",
 	"code", {
-		"Code/Script.lua"
+		"Code/Script.lua",
 	},
 	"lua_revision", LuaRevision or 244275,
 	"description", [[Shows the map before you load it in the select colony screen.
+Optionally show anomaly breakthroughs for that location.
+Omega breakthroughs are at best 1/3, planetary anomalies not included (maybe next update).
 
 For those wondering: Purple = mountainous, other colours = buildable.
 
-All maps: https://github.com/ChoGGi/SurvivingMars_CheatMods/tree/master/Mods%20ChoGGi/Map%20Images%20Pack/Maps]],
+View all maps: https://github.com/ChoGGi/SurvivingMars_CheatMods/tree/master/Mods%20ChoGGi/Map%20Images%20Pack/Maps]],
 })